shop_bot_backend/requirements.txt
langchain
langchain-google-genai
langchain-community
langchain-core
psycopg[binary]
pydantic
python-dotenv
fastapi
uvicorn